Implement logging that can get to ELK
This adds more logging into the front end, and also logs to a file so
ELK can get them.
All public calls to the CBS are logged with the return code.
This is not to say logging is completed in the CBS, more may come later
and under a more structured format.

Change-Id: I63c966876972e06c51276eaf1204b5b9b9e6cea7
Issue-ID: DCAEGEN2-387
Signed-off-by: Tommy Carpenter <tommy@research.att.com>