Update patch set 2
Patch Set 2:

@Bill this PR doesn't remove 10000. 10000 stays exactly the same (HTTP). It only adds HTTPS on 443. The magic is the ENV variable LISTEN_PORT that gets used by the parent Docker container to generate another NGINX.conf that you don't see here (the two coalesce together, they are "additive"). 

Privately there have been grumblings about running both HTTP and HTTPS simultaneously on different ports. However, I assert that this can be controlled via port mappings as the README says. When you deploy this, don't map 443 or 10000 if you don't want to. 

Unless the security credentials complain, we can roll with this for now. If something official says we must only run HTTP or HTTPS on one port within the code itself, this will need to be revisited.

Patch-set: 2