Update patch set 3
Patch Set 3:

Lusheng
This adds HTTPS but HTTP still runs on 10000 at the request of some others. So I only bumped the minor in setup.py. You should change the version in the setup.py to match 3 if you feel about that. But this doesn't change the original HTTP API.

Patch-set: 3
Label: Code-Review=0