Update patch set 1
Patch Set 1: Code-Review+1

The use of HTTP versus HTTPS seems like it's a deployment decision, rather than a fundamental part of the API's definition.  The set of requests that the API supports, the syntax of those requests, and the semantics (captured only informally in the Swagger/OpenAPI definition via descriptions) haven't changed.   You've changed the implementation of the config binding service to allow the deployer the option to use HTTPS, an option that wasn't available before.

You could make the argument the API version should be bumped because the API might now be served over HTTPS whereas before the code implementing the API did not permit HTTPS at all.  I don't find that compelling, but I don't strenuously object.

Patch-set: 1
Label: Code-Review=+1